Sort walker seed commits by time
This lets us init the walker's sorting method to GIT_SORT_TOPOLOGICAL instead of TOPOLOGICAL|TIME.

Previously, we couldn't do without TIME because the order of seed commits was unstable.

A pleasant side effect is that this yields a nicer-looking topo graph for TestGitRepository. This also pushes ugly "commit islands" that never merge into the main branch to the very end of the graph for repos such as pygit2.