Use deleteLater instead of WA_DeleteOnClose on context menus
Can't set WA_DeleteOnClose on context menus because this crashes on macOS e.g. when exporting a patch from a FileListView. (Context menu gets deleted while callback is running)

Qt docs say: "for the object to be deleted, the control must return to the event loop from which deleteLater() was called" -- I suppose this means the context menu won't be deleted until the FileList has control again.