RepoTaskRunner: Fix rare repo refresh miss if QThreadPool takes a while to flush
RepoTaskRunner's QThreadPool sometimes takes a little longer than usual to flush its worker thread, even though the actual worker function is done. RepoTaskRunner would then appear busy to the postTask callback.

This could cause RepoWidget to miss a critical post-task refresh - it postpones the refresh if RepoTaskRunner is busy. For example, after staging a file, the refresh could be missed, making the staging operation look like it had no effect.

This patch fixes this issue. We simply wait for the QThreadPool to be empty before firing postTask.

To reproduce the bug prior to this commit: In RepoTaskRunner._emitNextToken, add `QThread.msleep(300)` right after the `emit` statement. This simulates QThreadPool taking a while to wrap up.